services/recorder/requirements.txt
pika==1.3.2
opencv-python==4.8.1.78
pyyaml==6.0